Intcode: Store memory as i32

+++ src/intcode.rs
@@ -1,11 +1,11 @@
 pub struct Program {
     ip: usize,
-    memory: Vec<usize>,
+    memory: Vec<i32>,
     halted: bool,
 }
 
 impl Program {
-    pub fn new(initial_memory: &Vec<usize>) -> Program {
+    pub fn new(initial_memory: &Vec<i32>) -> Program {
         Program {
             ip: 0,
             memory: initial_memory.clone(),
@@ -19,11 +19,11 @@ impl Program {
         }
     }
 
-    pub fn output(&self) -> usize {
+    pub fn output(&self) -> i32 {
         self.value(0)
     }
 
-    fn value(&self, addr: usize) -> usize {
+    fn value(&self, addr: usize) -> i32 {
         self.memory[addr]
     }
 
@@ -37,7 +37,7 @@ impl Program {
         };
     }
 
-    fn stepval(&mut self) -> usize {
+    fn stepval(&mut self) -> i32 {
         let value = self.memory[self.ip];
         self.ip += 1;
         value
@@ -47,14 +47,14 @@ impl Program {
         let addr0 = self.stepval();
         let addr1 = self.stepval();
         let dst = self.stepval();
-        self.memory[dst] = self.value(addr0) + self.value(addr1);
+        self.memory[dst as usize] = self.value(addr0 as usize) + self.value(addr1 as usize);
     }
 
     fn mult(&mut self) {
         let addr0 = self.stepval();
         let addr1 = self.stepval();
         let dst = self.stepval();
-        self.memory[dst] = self.value(addr0) * self.value(addr1);
+        self.memory[dst as usize] = self.value(addr0 as usize) * self.value(addr1 as usize);
     }
 
     fn halt(&mut self) {
